Added build to scripts in package.json file: packages>uikit>package.json problem solved:
Found the issue! The @pancakeswap/uikit package doesn't have a build script, which is what Turbo is looking for. It only has build:rollup.
The problem is that Vercel is trying to build the uikit package as if it's a Next.js app, but it's actually a component library that uses Rollup for building.